登录

+++ login/src/main/java/com/zb/manage/biz/service/LessonService.java
@@ -1,6 +1,7 @@
 package com.zb.manage.biz.service;
 
 import com.zb.manage.dal.mapper.LessonMapper;
+import com.zb.manage.dal.mapper.SelectionMapper;
 import com.zb.manage.dal.model.Lesson;
 import com.zb.manage.dal.mapper.TeacherMapper;
 import com.zb.manage.dal.model.Teacher;
@@ -19,12 +20,16 @@ public class LessonService {
     @Resource
     private TeacherMapper teacherMapper;
 
+    @Resource
+    private SelectionMapper selectionMapper;
+
     public List<Lesson> findAll() {
         return lessonMapper.findAll();
     }
+
     public void add(BigInteger id, String lename, BigInteger lenumber, BigInteger teachid, String letime) {
         Teacher teacher = teacherMapper.getByTeachid(teachid);
-        if(null == teacher) {
+        if (null == teacher) {
             throw new IllegalArgumentException("教师ID不合法:" + teachid);
         }
         Lesson lesson = new Lesson();
@@ -38,10 +43,14 @@ public class LessonService {
     }
 
     public void delete(BigInteger id) {
+        int selectionCount = selectionMapper.countByLessonId(id);
+        if (selectionCount > 0) {
+            throw new IllegalArgumentException("选课表有依赖此课程，不得删除");
+        }
         lessonMapper.delete(id);
     }
 
-    public void update(Lesson lesson){
+    public void update(Lesson lesson) {
         lessonMapper.update(lesson);
     }
